add explicit code to get the HTTP Response on error
The version of Mithril we're using (1.1.3) doesn't properly pass through
the `xhr.status` on error so when we try to handle error conditions we
can't tell if it's a 401 or not. This caused us to log out on every
error, including 404s.

This adds some logic to our own request handler from a more recent
version of Mithril to copy over the HTTP Status code from the XHR
request so we can tell what it was and dispatch accordingly.